build: fix symlink windows git bug
When cloning on Windows, symlinks are replaced with text file containing
the path of the symlink, which causes Linux build to fail if the clone
was done on a Windows host.
Add a fix to all external_toolchain.cmake files so that it will only
treat this file if it is a symlink - we don't support cross compilation
when building on Windows.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>