framework: tlvf: Add hasTlv() to CmduMessage class
Preparative commit.
Add hasTlv() API to CmduMessage class which goes through the list of
TLVs and returns true if there is at least one TLV of the requested
type.

The use for this will be in parsing WSC CMDUs which may be originated
from either the controller or the agent, and the only difference which
allows diffrentiating between such CMDUs sent from the agent or the
controller are using the TLVs which are inside the CMDU.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>