controller: autoconfig: fix search response
The 1905 standard requires the autoconfig search response to include the
same MID (Message Identifier) as the autoconfig search message.
This was not the case in prplMesh controller implementation which made
the testbed Marvell CTT Agent to drop all autoconfig search responses,
so fix it to use same mid as in the received message.

Signed-off-by: Tomer Eliyahu <tomer.b.eliyahu@intel.com>