tools: beerocks_analyzer: remove sta_ids construction
Some parts of the file parser was trying to assign unique sta_ids in a
way that was way more complex than what it needs to be.

Also it relied on parent_bssid, which is not in the data anymore.

This commit removes the sta_ids, which will be reimplemented in a
cleaner way later when/if its needed.

Signed-off-by: Raphaël Mélotte <raphael.melotte@essensium.com>