tools: beerocks_analyzer: clear the graph when receiving the mark
beerocks_analyzer uses a mark to know when the connection map is
complete.
Previously, when a node disappeared from one update to the
other (without receiving a Disconnect event), it was kept in the graph.

This commit clears the graph when a new mark is read, to make sure
that the graph represents the new state of the map.

Signed-off-by: Raphaël Mélotte <raphael.melotte@essensium.com>