reorganize config data to fit qualcomm
DEBUG ONLY - do not merge

Using this we are able to run an agent only and correctly parse
Qualcomm controller M2.

- Changed the order of the config data so that network key comes after vendor extension
This is needed for debugging qualcomm issue since it sends it this way
and we don't have TLVF support for parsing attributes out of order yet.

- Changed vendor extension attribute (multiap_attr in configData) to include the Version2
attribute since this is what Qualcomm sends, although it is not clear
from the specification that this should be the case.
We do need to support more than one subelement cause according to the specification
there could be 1 or more subelements.
The solution should be to change sWscAttrVendorExtMultiAp to be a TLV class
with a variable length list of subelements.

- Change network key to be a variable length list instead of fixed length.
This should be the first fix to be merged to prplMesh officially - network key
is like SSID, it is not fixed size attribute.