framework: Fix update vendor-specific tlv length
The current situation is that when finalizing VENDOR_SPECIFIC_MESSAGE,
which includes only one vendor-specific tlv, the tlv length is being
updated. The problem is that if a vendor-specific tlv is being added to
another CMDU type, an update of the tlv length does not occur.

Fix this by going over all the classes, and if vendor-specific tlv is
identified, it updates the length of it by adding the length of the
the tlv data class, which is two classes after the vendor-specific tlv
class.

Remove update of vendor-specific tlvs lengths on controller and slave
which is no longer necessary.

Signed-off-by: Moran Shoeg <moranx.shoeg@intel.com>