tools: beerocks_analyzer: implement threads termination
Previously the application tried to call terminate() on the treads,
but that wasn't implemented.

This commit implements a specific terminate() method for each thread,
and also set threads to be daemon so that they are killed anyway if
the application exits.

Note that there is still a problem with the way the threading is done
in the application (resources like files are not always dealed with
correctly - use context managers) but it will mitigate risks of having
background threads that keeps running after the application exits.

Signed-off-by: Raphaël Mélotte <raphael.melotte@essensium.com>