controller: beerocks_cli: increase the periodic updates frequency
As of now the analyzer only receive periodic updates.

In addition to periodic updates, it would be best to catch nw_map
changes events to send update to the analyzer, but that does not work
at the moment.

As a temporary workaround, this commit increases the frequency of the
periodic updates to see changes on the analyzer more quickly.

Signed-off-by: Raphaël Mélotte <raphael.melotte@essensium.com>