tools: beerocks_analyzer: revive the SSH option
It was previously possible to connect to the controller via SSH, but
it did not work anymore.

This commit allows to usage the SSH option again.
It removes the usage of a password, which means that you need a
private key to be registered in your local ssh agent.
A future commit will explain how to do that.

Note that the current version of paramiko is not compatible with the
latest openssh private key format, so a temporary fork is
needed (included in the requirement.txt).

Signed-off-by: Raphaël Mélotte <raphael.melotte@essensium.com>